Fixed a warning in MSVC6
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@4837 d7cf8633-e32d-0410-b094-e92efae38249

+++ prvm_exec.c
@@ -391,7 +391,7 @@ void PRVM_ExecuteProgram (func_t fnum, const char *errormessage)
 	prvm_eval_t	*ptr;
 	int		profile, startprofile, cachedpr_trace, exitdepth;
 
-	if (!fnum || fnum >= prog->progs->numfunctions)
+	if (!fnum || fnum >= (unsigned int)prog->progs->numfunctions)
 	{
 		if (prog->self && PRVM_G_INT(prog->self->ofs))
 			PRVM_ED_Print(PRVM_PROG_TO_EDICT(PRVM_G_INT(prog->self->ofs)));
